Ext_Cisco_NXOS_Shell_Package/Resource Drivers - Python/Generic Cisco NXOS Driver Version3 Extended/requirements.txt
cloudshell-automation-api>=7.0,<7.1
cloudshell-networking-devices>=2.3,<2.4
cloudshell-networking-cisco>=5.2,<5.3
cloudshell-networking-cisco-nxos>=5.2,<5.3